# apps/web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Civitas API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/account/favourites/{urn}": {
      "put": {
        "tags": [
          "account"
        ],
        "summary": "Save Account Favourite",
        "operationId": "save_account_favourite_api_v1_account_favourites__urn__put",
        "parameters": [
          {
            "name": "urn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Urn"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSchoolStateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "account"
        ],
        "summary": "Remove Account Favourite",
        "operationId": "remove_account_favourite_api_v1_account_favourites__urn__delete",
        "parameters": [
          {
            "name": "urn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Urn"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedSchoolStateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SavedSchoolStateResponse": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "saved",
              "not_saved",
              "requires_auth",
              "locked"
            ],
            "title": "Status"
          },
          "saved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Saved At"
          },
          "capability_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capability Key"
          },
          "reason_code": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "free_baseline",
                  "premium_capability_missing",
                  "anonymous_user",
                  "artefact_not_published",
                  "artefact_not_supported",
                  "entitlement_expired",
                  "entitlement_revoked"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason Code"
          }
        },
        "type": "object",
        "required": [
          "status",
          "saved_at",
          "capability_key",
          "reason_code"
        ],
        "title": "SavedSchoolStateResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}